Menus in the levels were disappearing when you clicked anywhere in the panel. The menu in the levels only disappear when the menu button is clicked now. Also changed the location of home button for the intro levels to reside under the Drop Down panel so it will disappear with menu.
In Unity in each scene, the main camera drop down to the canvas drop down to the drop down button to the drop down panel. For drop down panel, the Raycast Target was unchecked in the Image(Scipt) section, and for all the text in the drop down panel, the Raycast Target was unchecked in the Text(Script) for each text piece in the panel. This fixed the menu disappearing when you clicked anywhere in panel.